🚚 Rename MainButton to ClickButton
